requirements.txt
llama-index==0.11.20
python-dotenv==1.0.1
chainlit==1.3.1